Initial commit
Implements generation and validation of SMT proofs, and JSON serialization. Adapted from the
github.com/confio/ics23-iavl project.

Note that because the lazyledger SMT hashes keys internally, we have to keep a mapping of preimages
ordered by hashed path, to determine which paths are adjacent to a non-existent key, and to test
path edge cases.